break-out RecordTypeFilters into dedicated component
