+++ src/components/Collections/RecordTypeFilters.js
@@ -1,0 +1,95 @@
+import React from 'react';
+import {
+  arrayOf, bool, func, shape, string,
+} from 'prop-types';
+import { connect } from 'react-redux';
+import {
+  Platform, StyleSheet, Switch, Text, View,
+} from 'react-native';
+
+import Colors from '../../constants/Colors';
+import { orderedResourceTypeFiltersSelector } from '../../redux/selectors';
+import { toggleResourceTypeFilter } from '../../redux/action-creators';
+
+const TypeFilterRow = ({
+  resourceType, label, filterOpen, toggleResourceTypeFilterAction,
+}) => {
+  let thumbColor;
+  if (Platform.OS === 'ios') {
+    thumbColor = 'white';
+  } else {
+    thumbColor = filterOpen ? Colors.primary : Colors.lightgrey;
+  }
+
+  return (
+    <View style={styles.categoryRow}>
+      <Text>{label}</Text>
+      <Switch
+        trackColor={{
+          false: Colors.mediumgrey,
+          true: Platform.OS === 'ios' ? Colors.primary : Colors.primaryLight,
+        }}
+        thumbColor={thumbColor}
+        onValueChange={() => toggleResourceTypeFilterAction(resourceType)}
+        value={filterOpen}
+      />
+    </View>
+  );
+};
+
+TypeFilterRow.propTypes = {
+  resourceType: string.isRequired,
+  label: string.isRequired,
+  filterOpen: bool.isRequired,
+  toggleResourceTypeFilterAction: func.isRequired,
+};
+
+const RecordTypeFilters = ({ orderedResourceTypeFilters, toggleResourceTypeFilterAction }) => (
+  <View>
+    <Text style={styles.drawerTitle}>Record Type Filters</Text>
+    {orderedResourceTypeFilters.map(({ type, typeIsEnabled, label }) => (
+      <TypeFilterRow
+        key={type}
+        resourceType={type}
+        label={label}
+        filterOpen={typeIsEnabled}
+        toggleResourceTypeFilterAction={toggleResourceTypeFilterAction}
+      />
+    ))}
+  </View>
+);
+
+RecordTypeFilters.propTypes = {
+  orderedResourceTypeFilters: arrayOf(shape({
+    type: string.isRequired,
+    typeIsEnabled: bool.isRequired,
+    label: string.isRequired,
+  })).isRequired,
+  toggleResourceTypeFilterAction: func.isRequired,
+};
+
+const mapStateToProps = (state) => ({
+  orderedResourceTypeFilters: orderedResourceTypeFiltersSelector(state),
+});
+
+const mapDispatchToProps = {
+  toggleResourceTypeFilterAction: toggleResourceTypeFilter,
+};
+
+export default connect(mapStateToProps, mapDispatchToProps)(RecordTypeFilters);
+
+const styles = StyleSheet.create({
+  drawerTitle: {
+    marginTop: 10,
+    marginBottom: 20,
+    fontSize: 20,
+    textAlign: 'center',
+  },
+  categoryRow: {
+    flexDirection: 'row',
+    justifyContent: 'space-between',
+    alignItems: 'center',
+    paddingHorizontal: 20,
+    marginVertical: 10,
+  },
+});
